make tests expandable like translator
