+++ packages/xlucene-translator/test/cases/queries/index.ts
@@ -1,0 +1,9 @@
+import type { TestCaseCollection } from './interfaces.js';
+import { arrays, empty_arrays } from './arrays.js';
+
+const tests: TestCaseCollection = {
+    arrays: arrays,
+    empty_arrays,
+};
+
+export default tests;
